+++ tests/tests_api_v1_1/test_api_v1_1_reset_password.py
@@ -85,6 +85,8 @@ class ResetPasswordTest(APITestCase):
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertEqual(response.data['message'], 'invalid data')
+        self.assertIn('_errors', response.data)
+        self.assertEqual(response.data['_errors'], ['INVALID_DATA'])
 
     def test_reset_password_fail_email(self):
 
@@ -96,6 +98,8 @@ class ResetPasswordTest(APITestCase):
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertEqual(response.data['message'], 'invalid data')
+        self.assertIn('_errors', response.data)
+        self.assertEqual(response.data['_errors'], ['INVALID_DATA'])
 
     def test_reset_password_fail_email_token_association(self):
         reset_password_url = '/api/v1.1/auth/reset-password/'
@@ -123,6 +127,8 @@ class ResetPasswordTest(APITestCase):
         )
         self.assertEqual(rsp.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertEqual(rsp.data['message'], 'invalid token')
+        self.assertIn('_errors', rsp.data)
+        self.assertEqual(rsp.data['_errors'], ['INVALID_TOKEN'])
 
     def test_reset_password_confirmation_wrong_serializers(self):
         reset_password_url = '/api/v1.1/auth/reset-password/'
@@ -175,6 +181,8 @@ class ResetPasswordTest(APITestCase):
         )
         self.assertEqual(rsp.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertEqual(rsp.data['message'], 'Invalid data')
+        self.assertIn('_errors', rsp.data)
+        self.assertEqual(rsp.data['_errors'], ['INVALID_DATA'])
 
     def test_reset_password_fail_wrong_token_in_confirmation(self):
         reset_password_url = '/api/v1.1/auth/reset-password/'
@@ -202,6 +210,8 @@ class ResetPasswordTest(APITestCase):
         )
         self.assertEqual(rsp.status_code, status.HTTP_400_BAD_REQUEST)
         self.assertEqual(rsp.data['message'], 'invalid token')
+        self.assertIn('_errors', rsp.data)
+        self.assertEqual(rsp.data['_errors'], ['INVALID_TOKEN'])
 
     def test_reset_password_fail_mismatch_passwords(self):
 
@@ -229,6 +239,8 @@ class ResetPasswordTest(APITestCase):
             },
         )
         self.assertEqual(rsp.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn('_errors', rsp.data)
+        self.assertEqual(rsp.data['_errors'], ['MISMATCH_PASSWORDS'])
 
     def test_reset_password_fail_token_too_old(self):
 
@@ -363,3 +375,5 @@ class ResetPasswordTest(APITestCase):
             },
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
+        self.assertIn('_errors', response.data)
+        self.assertEqual(response.data['_errors'], ['INVALID_PARAMETER'])