+++ tests/tests_api_v1_1/test_filters.py
@@ -24,7 +24,7 @@ from concrete_datastore.api.v1.datetime import format_datetime
 
 
 @override_settings(DEBUG=True)
-class FilterSupportingComparaisonBackendTestCase(APITestCase):
+class FilterSupportingComparisonBackendTestCase(APITestCase):
     def setUp(self):
         self.user = User.objects.create_user(
             'johndoe@netsach.org'
@@ -37,11 +37,7 @@ class FilterSupportingComparaisonBackendTestCase(APITestCase):
         UserConfirmation.objects.create(user=self.user, confirmed=True).save()
         url = '/api/v1.1/auth/login/'
         resp = self.client.post(
-            url,
-            {
-                "email": "johndoe@netsach.org",
-                "password": "plop",
-            },
+            url, {"email": "johndoe@netsach.org", "password": "plop"}
         )
         self.token = resp.data['token']
         for i in range(5):
@@ -84,11 +80,7 @@ class FilterSupportingOrBackendTestClass(APITestCase):
         UserConfirmation.objects.create(user=self.user, confirmed=True).save()
         url = '/api/v1.1/auth/login/'
         resp = self.client.post(
-            url,
-            {
-                "email": "johndoe@netsach.org",
-                "password": "plop",
-            },
+            url, {"email": "johndoe@netsach.org", "password": "plop"}
         )
         self.token = resp.data['token']
 
@@ -306,7 +298,8 @@ class FilterWithInvalidFields(APITestCase):
             {
                 'message': 'filter against {} is not allowed'.format(
                     requested_filter
-                )
+                ),
+                '_errors': ["INVALID_QUERY"],
             },
         )
 
@@ -322,7 +315,8 @@ class FilterWithInvalidFields(APITestCase):
             {
                 'message': 'filter against {} is not allowed'.format(
                     requested_filter
-                )
+                ),
+                '_errors': ["INVALID_QUERY"],
             },
         )
 
@@ -336,7 +330,8 @@ class FilterWithInvalidFields(APITestCase):
             {
                 'message': 'filter against {} is not allowed'.format(
                     requested_filter
-                )
+                ),
+                '_errors': ["INVALID_QUERY"],
             },
         )