+++ tests/tests_secure_token/test_retrieve_token.py
@@ -27,12 +27,7 @@ class RetrieveTokenTestCase(APITestCase):
         url = '/api/v1.1/secure-connect/retrieve-token/'
 
         # POST a valid email
-        resp = self.client.post(
-            url,
-            {
-                "email": 'usera@netsach.org'
-            },
-        )
+        resp = self.client.post(url, {"email": 'usera@netsach.org'},)
         self.assertEqual(
             resp.status_code, status.HTTP_201_CREATED, msg=resp.content
         )
@@ -48,12 +43,7 @@ class RetrieveTokenTestCase(APITestCase):
         url = '/api/v1.1/secure-connect/retrieve-token/'
 
         # POST a valid email
-        resp = self.client.post(
-            url,
-            {
-                "email": 'user42@netsach.org'
-            },
-        )
+        resp = self.client.post(url, {"email": 'user42@netsach.org'},)
         self.assertEqual(
             resp.status_code, status.HTTP_400_BAD_REQUEST, msg=resp.content
         )
@@ -70,20 +60,10 @@ class RetrieveTokenTestCase(APITestCase):
         url = '/api/v1.1/secure-connect/retrieve-token/'
 
         # POST a valid email
-        resp = self.client.post(
-            url,
-            {
-                "email": 'usera@netsach.org'
-            },
-        )
+        resp = self.client.post(url, {"email": 'usera@netsach.org'},)
         secure_tokens = SecureConnectToken.objects.filter(user=self.userA)
         self.assertEqual(secure_tokens.count(), 1)
-        resp = self.client.post(
-            url,
-            {
-                "email": 'usera@netsach.org'
-            },
-        )
+        resp = self.client.post(url, {"email": 'usera@netsach.org'},)
         self.assertEqual(
             resp.status_code,
             status.HTTP_429_TOO_MANY_REQUESTS,
@@ -98,12 +78,7 @@ class RetrieveTokenTestCase(APITestCase):
         url = '/api/v1.1/secure-connect/retrieve-token/'
 
         # POST a valid email
-        resp = self.client.post(
-            url,
-            {
-                "email": 'usera@netsach.org'
-            },
-        )
+        resp = self.client.post(url, {"email": 'usera@netsach.org'},)
         secure_tokens = SecureConnectToken.objects.filter(user=self.userA)
         self.assertEqual(secure_tokens.count(), 1)
         first_token = secure_tokens.first()
@@ -113,12 +88,7 @@ class RetrieveTokenTestCase(APITestCase):
         first_token.save()
 
         # Creation of second token
-        resp = self.client.post(
-            url,
-            {
-                "email": 'usera@netsach.org'
-            },
-        )
+        resp = self.client.post(url, {"email": 'usera@netsach.org'},)
         self.assertEqual(
             resp.status_code, status.HTTP_201_CREATED, msg=resp.content
         )
@@ -131,12 +101,9 @@ class RetrieveTokenTestCase(APITestCase):
         url = '/api/v1.1/secure-connect/retrieve-token/'
 
         # POST an invalid email
-        resp = self.client.post(
-            url,
-            {
-                "email": 'user42ànetsachcom'
-            },
-        )
+        resp = self.client.post(url, {"email": 'user42ànetsachcom'},)
         self.assertEqual(
             resp.status_code, status.HTTP_400_BAD_REQUEST, msg=resp.content
         )
+        self.assertIn('_errors', resp.data)
+        self.assertEqual(resp.data['_errors'], ['INVALID_DATA'])